requirements.txt
# ===========================================
# PiyP Backend - Python Dependencies
# ===========================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & ORM
supabase==2.3.0
postgrest==0.13.0

# Authentication & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.25.1
aiohttp==3.9.1

# PDF Processing (for RefMan)
pypdf==3.17.1
pdfplumber==0.10.3
python-magic==0.4.27

# File handling
python-magic==0.4.27

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# AI/ML (Optional - for later phases)
# openai==1.3.5
# anthropic==0.7.0
# langchain==0.0.340
# chromadb==0.4.18

# Utilities
tenacity==8.2.3
redis==5.0.1
